requirements.txt
django==2.1.5
django-ckeditor==5.6.1
django-mptt==0.9.1
django-sitetree==1.12.0
pillow==4.1.0
django-contrib-comments==1.9.0
scikit-learn==0.20.2
joblib==0.13.1
django-registration==3.0
exifread==2.1.2
django-location-field==2.1.0